LayoutTests/fast/dom/HTMLDataGridElement/DataGridColumns-basic.html

<html>
<body>
    <pre id="console"></pre>
    <datagrid id="test"></datagrid>
    <script>
        if (window.layoutTestController) {
            layoutTestController.dumpAsText();
            layoutTestController.waitUntilDone();
        }

        function log(msg)
        {
            document.getElementById('console').appendChild(document.createTextNode(msg + "\n"));
        }

        var datagrid = document.getElementById('test');

        // Test initial value.
        if (datagrid.columns != null) {
            log("PASS: Initial value of datagrid.columns is defined.");
        } else {
            log("FAIL: Initial value of datagrid.columns was null.");
        }

        // Test initial length.
        if (datagrid.columns.length == 0) {
            log("PASS: Initial length of datagrid.columns is 0.");
        } else {
            log("FAIL: Initial length of datagrid.columns was " + datagrid.columns.length + ".");
        }
        
        // Test adding a column.
        var column = datagrid.columns.add("fromcol", "From", "text", true, 2);
        
        // Test return value from adding the column.
        if (column != null) {
            log("PASS: Column returned from add.");
        } else {
            log("FAIL: No column returned.");
        }
        
        // Make sure the column list now has one column.
        if (datagrid.columns.length == 1) {
            log("PASS: We have one column in the list.");
        } else {
            log("FAIL: Expected one column in the list but have " + datagrid.columns.length + " columns instead.");
        }
        
        // Make sure the column we got back from add is the one in the list.
        if (datagrid.columns[0] == column) {
            log("PASS: The column at index 0 is the column we got back from add.");
        } else {
            log("FAIL: The column at index 0 is not the one we expected.");
        }
        
        if (datagrid.columns["fromcol"] == column) {
            log("PASS: The column when referenced by name is the one we got back from add.");
        } else {
            log("FAIL: The column when referenced by name is not the one we expected.");
        }

        // Examine the column's properties to make sure they are what we expected.
        if (column.id == "fromcol") {
            log("PASS: The column's ID is fromcol as expected.");
        } else {
            log("FAIL: The column's ID should be fromcol but is " + column.id + " instead.");
        }
        
        if (column.label == "From") {
            log("PASS: The column's label is From as expected.");
        } else {
            log("FAIL: The column's label should be From but is " + column.label + " instead.");
        }

        if (column.type == "text") {
            log("PASS: The column's type is text as expected.");
        } else {
            log("FAIL: The column's type should be text but is " + column.type + " instead.");
        }
    
        if (column.primary) {
            log("PASS: The column's primary property is true as expected.");
        } else {
            log("FAIL: The column's primary property should be true but isn't.");
        }

        if (column.sortable == 2) {
            log("PASS: The column's sortable property is 2 as expected.");
        } else {
            log("FAIL: The column's sortable property should be 2 but is " + column.sortable + " instead.");
        }

        // Make sure we got set as the primary column.
        if (datagrid.columns.primaryColumn == column) {
            log("PASS: The column returned from add is now the primary column.");
        } else {
            log("FAIL: The column is not the primary column as expected.");
        }

        if (window.layoutTestController)
            layoutTestController.notifyDone();
    </script>
</body>
</html>
